Update auth&acces and role&permission logic

See merge request mrs/unity/document-manager!90

+++ DM/DM/Controllers/OrganizationController.cs
@@ -1,9 +1,15 @@
 ﻿using System.Threading.Tasks;
 
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 
 using DM.Domain.Interfaces;
 using DM.Domain.Models;
+using DM.Domain.Implementations;
+using DM.Domain.Helpers;
+
+using DM.DAL;
+using DM.DAL.Entities;
 
 using DM.Helpers;
 
@@ -13,34 +19,55 @@ namespace DM.Controllers
     [Route("api/organization")]
     public class OrganizationController : ControllerBase
     {
-        private readonly IOrganizationService _organizationService;
+        private readonly DmDbContext _context;
+        private readonly UserEntity _currentUser;
 
-        public OrganizationController(IOrganizationService organizationService)
+        private readonly IOrganizationService _organizationService;
+        private readonly ILogger<OrganizationService> _logger;
+
+        public OrganizationController(DmDbContext context, CurrentUserService currentUserService, IOrganizationService organizationService, ILogger<OrganizationService> logger)
         {
+            _context = context;
+            _currentUser = currentUserService.CurrentUser;
             _organizationService = organizationService;
+            _logger = logger;
         }
 
-        [Authorize(new string[] { RoleConst.Admin, RoleConst.Owner })]
         [HttpGet]
+        [Authorize]
         public async Task<IActionResult> GetAll()
         {
+
+            var permission = AuthorizationHelper.CheckUserPermissionsForRead(_context, _currentUser, PermissionType.Organization);
+
+            if (!permission) return StatusCode(403);
+
             var organizations = await _organizationService.GetAll();
 
             return Ok(organizations);
         }
 
-        [Authorize(new string[] { RoleConst.Admin, RoleConst.Owner })]
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Create(OrganizationModelForCreate organizationModel)
         {
+            var permission = AuthorizationHelper.CheckUserPermissionsForCreate(_context, _currentUser, PermissionType.Organization);
+
+            if (!permission) return StatusCode(403);
+
             var checker = await _organizationService.Create(organizationModel);
 
             return Ok(checker);
         }
 
         [HttpPut]
+        [Authorize]
         public async Task<IActionResult> Update(OrganizationModelForUpdate organizationModel)
         {
+            var permission = AuthorizationHelper.CheckUserPermissionsForUpdate(_context, _currentUser, PermissionType.Organization);
+
+            if (!permission) return StatusCode(403);
+
             var checker = await _organizationService.Update(organizationModel);
 
             if (checker == false) return BadRequest("No such organization exists");
@@ -48,10 +75,14 @@ namespace DM.Controllers
             return Ok(checker);
         }
 
-        [Authorize(new string[] { RoleConst.Admin, RoleConst.Owner })]
         [HttpDelete]
+        [Authorize]
         public async Task<IActionResult> Delete(long organizationId)
         {
+            var permission = AuthorizationHelper.CheckUserPermissionsForDelete(_context, _currentUser, PermissionType.Organization);
+
+            if (!permission) return StatusCode(403);
+
             var checker = await _organizationService.Delete(organizationId);
 
             if (checker == false)
